tm/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.9
django-bootstrap-v5==1.0.7
django-rest-framework==0.1.0
djangorestframework==3.12.4
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
